refactor(lowy): move isAttentionState into terminal/ + collapse agentBucket duplication
`isAttentionState` lived in `ui/agentDisplay.ts` (a display-strings module) but
was consumed by domain code — `staleness.ts` for the attention-state
exemption, `useTerminalAlerts.ts` for the badge gate. Domain importing from
display inverted the dependency arrow; this commit moves the predicate to
its own focused module at `terminal/agentState.ts` and re-exports from
`agentDisplay.ts` for backward compatibility.

While there, the equivalence class duplication flagged by the lowy
cross-validation gets closed: `agentBucket` (`dockModel.ts`) was repeating
the literal `waiting | awaiting_user` and `thinking | tool_use` lists in
its `match` block — now it consumes `isAttentionState` / `isWorkingState`
from `agentState.ts`. Exhaustiveness preserved via a `state satisfies never`
fence at the fall-through, so any new state literal added to
`AgentInfo["state"]` will compile-fail at the bucket site until it joins
one of the two predicates.
